Split ic checking from customer form, allow subs to be created w/o existing customer
